towards #520; refactored to compartmentalize vue frontend engine
