+++ jpcore/justpy_config.py
@@ -53,3 +53,5 @@ AGGRID = config("AGGRID", cast=bool, default=True)
 AGGRID_ENTERPRISE = config("AGGRID_ENTERPRISE", cast=bool, default=False)
 
 NO_INTERNET = config("NO_INTERNET", cast=bool, default=True)
+
+COMPONENT_ENGINE_TYPE = config("COMPONENT_ENGINE_TYPE", cast=str, default="vue")
