+++ jpcore/template.py
@@ -25,6 +25,7 @@ class Context:
         self.page_options = PageOptions(context_dict.get("page_options", {}))
         self.use_websockets_js=self.context_dict.get("use_websockets","true")
         self.page_id_js=context_dict["page_id"]
+        self.component_engine_type = context_dict["component_engine_type"]
         self.title_js=self.get_js_option("title", "JustPy")
         self.redirect_js=self.get_js_option("redirect","")
         self.display_url_js=self.get_js_option("display_url","")
@@ -59,7 +60,8 @@ class Context:
         """
         generate the html lines for justpy to work 
         """
-        html=self.as_script_src("justpy_core")
+        html=self.as_script_src(f"component_generator", subdir=self.component_engine_type)
+        html+=self.as_script_src("justpy_core")
         html+=f"""{indent}<script>
 {indent}  var page_id = {self.page_id_js};
 {indent}  var use_websockets = {self.use_websockets_js};
@@ -70,8 +72,9 @@ class Context:
         html+=f"{indent}<script>\n{self.as_javascript_setup(indent)}\n{indent}</script>\n"
         return html
     
-    def as_script_src(self,file_name:str,indent:str="  "):
-        src= f"{indent}<script src='/templates/js/{file_name}.js'></script>\n"
+    def as_script_src(self,file_name:str, indent:str="  ", subdir=""):
+        
+        src= f"{indent}<script src='/templates/js/{subdir}/{file_name}.js'></script>\n"
         return src
     
     def as_script_srcs(self,indent:str="  "):
@@ -79,8 +82,9 @@ class Context:
         generate a list of javascript files to be imported
         """
         srcs = ""
+        srcs +=self.as_script_src("event_handler", indent)
+        #TODO: read from component dir 
         for file_name in [
-            "event_handler",
             "html_component",
             "quasar_component",
             "chartjp",
@@ -93,7 +97,7 @@ class Context:
             "katexjp",
             "editorjp",
         ]:
-            srcs +=self.as_script_src(file_name,indent)
+            srcs +=self.as_script_src(file_name, indent, subdir=self.component_engine_type)
         return srcs
     
     def as_javascript_setup(self,indent):
