+++ justpy/templates/js/justpy_core.js
@@ -156,33 +156,10 @@ class JustpyCore {
         } else {
             this.setupNoWebSocket();
         }
-        this.createApp1();
-        this.registerAllEvents()
-    }
-
-    /**
-     * create the global app1 Vue object
-     */
-    createApp1() {
-        app1 = new Vue({
-            el: '#components',
-            data: {
-                justpyComponents: justpyComponents
-            },
-            render: function (h) {
-                var comps = [];
-                for (var i = 0; i < this.justpyComponents.length; i++) {
-                    if (this.justpyComponents[i].show) {
-                        comps.push(h(this.justpyComponents[i].vue_type, {
-                            props: {
-                                jp_props: this.justpyComponents[i]
-                            }
-                        }))
-                    }
-                }
-                return h('div', {}, comps);
-            }
-        });
+       app1 = createApp();
+      this.registerAllEvents();
+      
+      
     }
 
     /**
