+++ justpy/templates/js/vue/bokehjp.js
