+++ justpy/templates/js/vue/html_component.js
