+++ justpy/templates/js/vue/katexjp.js
