+++ justpy/templates/js/vue/plotlyjp.js
