+++ justpy/templates/js/vue/quasar_component.js
